backend/requirements.txt
attrs==25.3.0
bandit==1.8.6
blinker==1.9.0
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
Flask==3.1.1
fonttools==4.59.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
networkx==3.5
numpy==2.3.1
packaging==25.0
pbr==6.1.1
pillow==11.3.0
pluggy==1.6.0
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.36.2
rich==14.0.0
rpds-py==0.26.0
setuptools==80.9.0
six==1.17.0
stevedore==5.4.1
typing_extensions==4.14.1
Werkzeug==3.1.3
